\name{arplsbaseline-package}
\alias{arplsbaseline-package}
\alias{arplsbaseline}
\docType{package}
\title{
  \packageTitle{arplsbaseline}
}
\description{
  \packageDescription{arplsbaseline}
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{ \packageAuthor{arplsbaseline}

Maintainer: \packageMaintainer{arplsbaseline}
}
\references{
  Baek, S.-J., Park, A., Ahn, Y.-J., & Choo, J. (2015).
  Baseline correction using asymmetrically reweighted penalized least squares smoothing.
  The Analyst, 140(1), 250–257. \href{https://doi.org/10.1039/c4an01061b}{Royal Society of Chemistry}
}
\keyword{ package }

\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{arplsbaseline}}
\section{Documentation Index}{\packageIndices{arplsbaseline}}