% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectrum.R
\name{as.data.frame.Spectrum}
\alias{as.data.frame.Spectrum}
\title{Coerce Spectrum object to data.frame}
\usage{
\method{as.data.frame}{Spectrum}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A Spectrum object.}

\item{row.names}{Not used. Included for consistency with \code{\link[base]{as.data.frame}}.}

\item{optional}{Not used. Included for consistency with \code{\link[base]{as.data.frame}}.}

\item{...}{Not used. Included for consistency with \code{\link[base]{as.data.frame}}.}
}
\value{
A data.frame with spectrum data.
}
\description{
Converts a Spectrum object into a data.frame.
}
\examples{
# example code to convert to dataframe
spec <- baseline(strawberry)
as.data.frame(spec)
}
