% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ham}
\alias{ham}
\title{Mid-Infrared Spectrum of Ham}
\format{
A dataframe with 2489 observations on 2 variables:
\describe{
  \item{Wavenumbers}{numeric Wavenumbers (1/cm)}
  \item{Ham}{numeric Signal Intensity}
  }
}
\source{
\href{https://github.com/QIBChemometrics/FTIR-food-library/blob/main/FTIR_Library_Spectra.csv}{QIB Chemometrics}
}
\usage{
data(ham)
}
\description{
Fingerprint region of Mid-Infrared Spectrum of cooked British Ham, 
using FTIR spectroscopy with ATR sampling (diamond).
}
\keyword{datasets}
