% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectrum.R
\name{plot.Spectrum}
\alias{plot.Spectrum}
\title{Spectra Visualisation}
\usage{
\method{plot}{Spectrum}(x, y = NULL, ...)
}
\arguments{
\item{x}{A Spectrum object.}

\item{y}{Not used. Included for consistency with \code{\link[graphics]{plot}}.}

\item{...}{Not used. Included for consistency with \code{\link[graphics]{plot}}.}
}
\description{
Plots the baseline, original data and corrected spectrum all in one plot using ggplot2.
}
\examples{
# example code
spec <- baseline(strawberry)
plot(spec)
}
