% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectrum.R
\name{print.Spectrum}
\alias{print.Spectrum}
\title{Print Spectrum}
\usage{
\method{print}{Spectrum}(x, ...)
}
\arguments{
\item{x}{A Spectrum object.}

\item{...}{Not used. Included for consistency with \code{\link[base]{print}}.}
}
\description{
Prints information about Spectrum object structure
}
\examples{
# example code to print
spec <- baseline(strawberry)
print(spec)
}
