% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{redwine}
\alias{redwine}
\title{Mid-Infrared Spectrum of Red Wine}
\format{
A dataframe with 235 observations on 2 variables.
\describe{
  \item{Wavenumbers}{numeric Wavenumbers (1/cm)}
  \item{Wine}{numeric Signal Intensity}
  }
}
\source{
\href{https://github.com/QIBChemometrics/FTIR-food-library/blob/main/FTIR_Library_Spectra.csv}{QIB Chemomentrics}
}
\usage{
redwine
}
\description{
Fingerprint region of Mid-Infrared Spectrum of Australian Cabernet Sauvignon Red Wine, 
using FTIR spectroscopy with ATR sampling (zinc selenide).
}
\references{
Kemsley EK, Defernez M, Marini, F (2019). 
Multivariate statistics: Considerations and confidences in food authenticity problems. 
Food Control, 105, 102-112.
}
\keyword{datasets}
