% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{strawberry}
\alias{strawberry}
\title{Mid-Infrared Spectrum of Strawberry Puree}
\format{
A dataframe with 235 observations on 2 variables.
\describe{
  \item{Wavenumbers}{numeric Wavenumbers (1/cm)}
  \item{Strawberry}{numeric Signal Intensity}
  }
}
\source{
\href{https://csr.quadram.ac.uk/example-datasets-for-download/}{QIB Chemometrics}
}
\usage{
strawberry
}
\description{
Mid-Infrared Spectrum of Strawberry Puree, 
using FTIR spectroscopy with ATR sampling.
}
\references{
Holland JK, Kemsley EK, Wilson RH. (1998). Use of Fourier transform 
 infrared spectroscopy and partial least squares regression for the detection of 
 adulteration of strawberry purees. Journal of the Science of Food and Agriculture, 76, 263-269.
}
\keyword{datasets}
