% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectrum.R
\name{summary.Spectrum}
\alias{summary.Spectrum}
\title{Spectrum Summary Statistics}
\usage{
\method{summary}{Spectrum}(object, ...)
}
\arguments{
\item{object}{A Spectrum object.}

\item{...}{Not used. Included for consistency with \code{\link[base]{summary}}.}
}
\description{
Prints summary statistics about the corrected spectrum. This calculates mean,
 median, standard deviation, minimum, maximum, signal range and skewness. The maximum also tells
 you at what wavenumber this occurs at.
}
\examples{
# example code
spec <- baseline(strawberry)
summary(spec)
}
