% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline}
\alias{baseline}
\title{baseline}
\usage{
baseline(data, lambda = 10000, start_mask = 0, end_mask = 0)
}
\arguments{
\item{data}{Dataframe of wavenumbers with corresponding signal intensity, first column must contain wavenumbers and second column signal intensity.}

\item{lambda}{Regularisation term to control the smoothness of the baseline, default is 1e4, common values to use are in powers of 10, from 1e2 to 1e8. Larger values give smoother baselines.}

\item{start_mask}{Lower limit of wavenumbers to ignore in baseline algorithm}

\item{end_mask}{Upper limit of wavenumbers to ignore in baseline algorithm}
}
\value{
Spectrum Object of custom Spectrum class
}
\description{
Calculate baseline using the ARPLS method by sourcing an Rcpp function
}
\examples{
# example code
# create spectrum from built-in dataset
baseline(ham, lambda = 1e6)
# can leave out lambda to use default
baseline(strawberry)
# masking example
baseline(redwine, start_mask = 1400, end_mask = 1500)
}
