% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline}
\alias{baseline}
\title{baseline}
\usage{
baseline(data, lambda = 10000)
}
\arguments{
\item{data}{Dataframe of frequencies/wavenumbers with signal intensity}

\item{lambda}{Regularisation term to control the smoothness of the baseline, default is 1e4}
}
\value{
Spectrum Object of custom Spectrum class
}
\description{
Calculate baseline using the ARPLS method by sourcing an Rcpp function
}
