% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectrum.R
\name{baseline_gam}
\alias{baseline_gam}
\title{Compute a GAM model to fit the computed baseline}
\usage{
baseline_gam(spectrum, full_summary = FALSE, return_gam = FALSE)
}
\arguments{
\item{spectrum}{Spectrum object}

\item{full_summary}{Boolean to specify whether the entire GAM model summary should be printed, default is FALSE}

\item{return_gam}{Boolean to specify whether to return the fitted GAM model, default is FALSE.}
}
\description{
Computes a GAM model to fit the computed baseline, providing summary and both visualisations of fit and residuals.
}
\examples{
# example code
spec <- baseline(ham)
baseline_gam(spec)
}
